pyproject.toml
[project]
name = "fnce-research"
version = "0.1.0"
description = "WRDS-based empirical finance research toolkit"
requires-python = ">=3.11"
dependencies = [
    "wrds",
    "pandas>=2.1",
    "numpy>=1.26",
    "pyarrow>=15",
    "scipy>=1.11",
    "statsmodels>=0.14",
    "linearmodels>=5.3",
    "python-dotenv>=1.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "openpyxl>=3.1",
    "jupyter>=1.0",
    "ipykernel>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov",
    "black",
    "ruff",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "W"]
